+++ cmd/psweb/main.go
@@ -2040,10 +2040,16 @@ func convertPeersToHTMLTable(peers []*peerswaprpc.PeerSwapPeer, allowlistedPeers
 
 			if stats.FeeSat > 0 {
 				flowText += "\nRevenue: +" + formatWithThousandSeparators(stats.FeeSat)
+				if stats.RoutedOut > 0 {
+					flowText += "\nRevenue PPM: " + formatWithThousandSeparators(stats.FeeSat*1_000_000/stats.RoutedOut)
+				}
 			}
 
 			if stats.PaidCost > 0 {
 				flowText += "\nCosts: -" + formatWithThousandSeparators(stats.PaidCost)
+				if stats.PaidOut > 0 {
+					flowText += "\nCosts PPM: " + formatWithThousandSeparators(stats.PaidCost*1_000_000/stats.PaidOut)
+				}
 			}
 
 			tooltip += flowText