+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.4.7
 
 - Remove resthost from peerswap.conf
+- Fix panic in appendInvoices
+- Add revenue and costs to tooltips
 
 ## 1.4.6